Fix native Carbon parameter sizing and refine macOS safety checks
Use the SDK ByteCount type on Intel and Apple silicon; document transparent overlay startup and add a regression for invalid saved preferences.

+++ macOS/Sources/PaperShadeMac/HotKeyController.swift
@@ -8,7 +8,7 @@ private let paperShadeHotKeyHandler: EventHandlerUPP = { _, event, context in
     var identifier = EventHotKeyID()
     let status = GetEventParameter(
         event, EventParamName(kEventParamDirectObject), EventParamType(typeEventHotKeyID),
-        nil, UInt32(MemoryLayout<EventHotKeyID>.size), nil, &identifier
+        nil, MemoryLayout<EventHotKeyID>.size, nil, &identifier
     )
     guard status == noErr, identifier.signature == paperShadeHotKeySignature else {
         return OSStatus(eventNotHandledErr)
